+++ lib/blocs/fs_entry_move/fs_entry_move_cubit.dart
@@ -77,7 +77,7 @@ class FsEntryMoveCubit extends Cubit<FsEntryMoveState> {
             folder.asEntity(), profile.wallet, driveKey);
 
         await _arweave.postTx(folderTx);
-        await _driveDao.updateFolder(folder);
+        await _driveDao.writeToFolder(folder);
       });
 
       emit(FolderEntryMoveSuccess());
@@ -94,7 +94,7 @@ class FsEntryMoveCubit extends Cubit<FsEntryMoveState> {
             file.asEntity(), profile.wallet, driveKey);
 
         await _arweave.postTx(fileTx);
-        await _driveDao.updateFile(file);
+        await _driveDao.writeToFile(file);
       });
 
       emit(FileEntryMoveSuccess());
