+++ lib/blocs/fs_entry_rename/fs_entry_rename_cubit.dart
@@ -65,7 +65,7 @@ class FsEntryRenameCubit extends Cubit<FsEntryRenameState> {
             folder.asEntity(), profile.wallet, driveKey);
 
         await _arweave.postTx(folderTx);
-        await _driveDao.updateFolder(folder);
+        await _driveDao.writeToFolder(folder);
       });
 
       emit(FolderEntryRenameSuccess());
@@ -80,7 +80,7 @@ class FsEntryRenameCubit extends Cubit<FsEntryRenameState> {
             file.asEntity(), profile.wallet, driveKey);
 
         await _arweave.postTx(fileTx);
-        await _driveDao.updateFile(file);
+        await _driveDao.writeToFile(file);
       });
 
       emit(FileEntryRenameSuccess());
