+++ lib/models/daos/drives_dao.dart
@@ -136,185 +136,6 @@ class DrivesDao extends DatabaseAccessor<Database> with _$DrivesDaoMixin {
       keyIv: Value(iv.bytes),
     );
   }
-
-  Future<void> applyEntityHistory(
-          String driveId, DriveEntityHistory entityHistory) =>
-      transaction(() async {
-        final drive = await (select(drives)..where((d) => d.id.equals(driveId)))
-            .getSingle();
-
-        // Create maps of folders and files whose contents were updated on the network, keyed by their ids.
-        final updatedFolders = <String, FolderEntriesCompanion>{};
-        final updatedFiles = <String, FileEntriesCompanion>{};
-
-        // Iterate through the history in reverse order to get the latest entity data we can write in.
-        for (final block in entityHistory.blockHistory.reversed) {
-          for (final entity in block.entities.reversed) {
-            // Ignore the entity if it did not come from the drive owner.
-            if (drive.ownerAddress != entity.ownerAddress) continue;
-
-            if (entity is FolderEntity) {
-              if (updatedFolders.containsKey(entity.id)) continue;
-
-              updatedFolders[entity.id] = FolderEntriesCompanion.insert(
-                id: entity.id,
-                driveId: entity.driveId,
-                parentFolderId: Value(entity.parentFolderId),
-                name: entity.name,
-                path: '',
-                dateCreated: Value(entity.commitTime),
-                lastUpdated: Value(entity.commitTime),
-              );
-            } else if (entity is FileEntity) {
-              if (updatedFiles.containsKey(entity.id)) continue;
-
-              updatedFiles[entity.id] = FileEntriesCompanion.insert(
-                id: entity.id,
-                driveId: entity.driveId,
-                parentFolderId: entity.parentFolderId,
-                name: entity.name,
-                dataTxId: entity.dataTxId,
-                size: entity.size,
-                path: '',
-                dateCreated: Value(entity.commitTime),
-                lastUpdated: Value(entity.commitTime),
-                lastModifiedDate: entity.lastModifiedDate,
-              );
-            }
-          }
-        }
-
-        // Write in the new folder and file contents without their paths
-        // and don't update the `dateCreated` if the entry already exists.
-        await batch((b) {
-          for (final updatedFolder in updatedFolders.values) {
-            b.insert(
-              folderEntries,
-              updatedFolder,
-              onConflict:
-                  DoUpdate((_) => updatedFolder.copyWith(dateCreated: null)),
-            );
-          }
-          for (final updatedFile in updatedFiles.values) {
-            b.insert(
-              fileEntries,
-              updatedFile,
-              onConflict:
-                  DoUpdate((_) => updatedFile.copyWith(dateCreated: null)),
-            );
-          }
-        });
-
-        // Construct a tree of the updated folders and files for path generation.
-        final staleFolderTree = <StaleFolderNode>[];
-
-        // Constructs a tree of folders and files that are children of the specified folder.
-        Future<StaleFolderNode> getStaleFolderTree(
-            FolderEntriesCompanion folder) async {
-          final folderId = folder.id.value;
-
-          // Get all the subfolders and files of this folder that are now stale.
-          final staleSubfolders = await (select(folderEntries)
-                ..where((f) => f.parentFolderId.equals(folderId)))
-              .get();
-          final staleFolderFiles = await (select(fileEntries)
-                ..where((f) => f.parentFolderId.equals(folderId)))
-              .get();
-          final staleFolderFilesMap = {
-            for (var f in staleFolderFiles) f.id: f.name
-          };
-
-          return StaleFolderNode(
-            folder,
-            // Recursively get the stale contents of this folder's subfolders.
-            await Future.wait(
-              staleSubfolders.map(
-                (f) => getStaleFolderTree(
-                    updatedFolders[f.id] ?? f.toCompanion(true)),
-              ),
-            ),
-            staleFolderFilesMap,
-          );
-        }
-
-        for (final folder in updatedFolders.values) {
-          final tree = await getStaleFolderTree(folder);
-
-          var newTreeIsSubsetOfExisting = false;
-          var newTreeIsSupersetOfExisting = false;
-          for (final existingTree in staleFolderTree) {
-            if (existingTree.searchForFolder(tree.folder.id.value) != null) {
-              newTreeIsSubsetOfExisting = true;
-            } else if (tree.searchForFolder(existingTree.folder.id.value) !=
-                null) {
-              staleFolderTree.remove(existingTree);
-              staleFolderTree.add(tree);
-              newTreeIsSupersetOfExisting = true;
-            }
-          }
-
-          if (!newTreeIsSubsetOfExisting && !newTreeIsSupersetOfExisting) {
-            staleFolderTree.add(tree);
-          }
-        }
-
-        Future<void> updateFolderTree(
-            StaleFolderNode node, String parentPath) async {
-          // If this is the root folder, we should not include its name as part of the path.
-          final folderPath = node.folder.parentFolderId.value != null
-              ? parentPath + '/' + node.folder.name.value
-              : '';
-
-          await (update(folderEntries)
-                ..where((f) => f.id.equals(node.folder.id.value)))
-              .write(FolderEntriesCompanion(path: Value(folderPath)));
-
-          for (final staleFileId in node.files.keys) {
-            final filePath = folderPath + '/' + node.files[staleFileId];
-            await (update(fileEntries)..where((f) => f.id.equals(staleFileId)))
-                .write(FileEntriesCompanion(path: Value(filePath)));
-          }
-
-          for (final staleFolder in node.subfolders) {
-            await updateFolderTree(staleFolder, folderPath);
-          }
-        }
-
-        for (final treeRoot in staleFolderTree) {
-          // Get the path of this folder's parent.
-          String parentPath;
-          if (treeRoot.folder.parentFolderId.value != null) {
-            parentPath = '';
-          } else {
-            parentPath = await (select(folderEntries)
-                  ..where(
-                      (f) => f.id.equals(treeRoot.folder.parentFolderId.value)))
-                .map((f) => f.path)
-                .getSingle();
-          }
-
-          await updateFolderTree(treeRoot, parentPath);
-        }
-
-        // Update paths of files whose parent folders were not updated.
-        final staleOrphanFiles = updatedFiles.values
-            .where((f) => updatedFolders[f.parentFolderId] == null);
-        for (final staleOrphanFile in staleOrphanFiles) {
-          final parentPath = await (select(folderEntries)
-                ..where(
-                    (f) => f.id.equals(staleOrphanFile.parentFolderId.value)))
-              .map((f) => f.path)
-              .getSingle();
-
-          await (update(fileEntries)..whereSamePrimaryKey(staleOrphanFile))
-              .write(FileEntriesCompanion(
-                  path: Value(parentPath + '/' + staleOrphanFile.name.value)));
-        }
-
-        await (update(drives)..whereSamePrimaryKey(drive)).write(
-            DrivesCompanion(
-                latestSyncedBlock: Value(entityHistory.latestBlockHeight)));
-      });
 }
 
 class CreateDriveResult {
@@ -323,22 +144,4 @@ class CreateDriveResult {
   final SecretKey driveKey;
 
   CreateDriveResult(this.driveId, this.rootFolderId, this.driveKey);
-}
-
-class StaleFolderNode {
-  final FolderEntriesCompanion folder;
-  final List<StaleFolderNode> subfolders;
-  final Map<String, String> files;
-
-  StaleFolderNode(this.folder, this.subfolders, this.files);
-
-  StaleFolderNode searchForFolder(String folderId) {
-    if (folder.id.value == folderId) return this;
-
-    for (final subfolder in subfolders) {
-      return subfolder.searchForFolder(folderId);
-    }
-
-    return null;
-  }
 }
